refactor: move err from entities to usecase
