+++ app/repository/pg/user.go
@@ -7,6 +7,7 @@ import (
 	"database/sql"
 	"fmt"
 	"github.com/go-devs-ua/octagon/app/entities"
+	"github.com/go-devs-ua/octagon/app/usecase"
 	"github.com/go-devs-ua/octagon/pkg/hash"
 	_ "github.com/lib/pq" // Standart blanc import for pq.
 	"strings"
@@ -35,7 +36,7 @@ func (r Repo) Add(user entities.User) (string, error) {
 
 	if err := r.DB.QueryRow(sqlStatement, user.FirstName, user.LastName, user.Email, hash.SHA256(user.Password)).Scan(&id); err != nil {
 		if strings.Contains(err.Error(), "unique_user_email") {
-			return "", entities.ErrDuplicateEmail
+			return "", usecase.ErrDuplicateEmail
 		}
 
 		return "", fmt.Errorf("error inserting into database: %w", err)
