+++ app/usecase/errors.go
@@ -1,4 +1,4 @@
-package entities
+package usecase
 
 import "errors"
